left-sidebar: Add back link underlines on hover, just for topics and PMs.
These were removed in 12f3fb599, along with the stream names and
right-sidebar user names mentioned in the commit message.  But unlike
those, these don't get any background change on hover, so we need some
way to show the hover.

We can probably still do something that looks nicer than this, but
it's better than not having it.