docs: Fix explanation of Debian LXC installation.
This didn't really make much sense - it presented a quickstart summary
of the instructions as an item in the list of differences from the
Ubuntu case, plus that quickstart only incorporated one of the two
actual differences so it didn't work anyway.  Given the networking
stuff, an actual quickstart isn't in the cards, so just cut that.

Also fix some small things while we're here.