hotkeys: Remove call to activity.blur_search().
The blur_search() function was removed in this commit:

See 812e296124a8b126f2b3f2c56afcade21d5f25bc

We now no longer attempt to call it.  It's not completely clear
to me what this did before, but we are rewriting a lot of the
keyboard navigation for search anyway.