hotkeys: Fix very recent regression with down key.
In 2f912dee3e8874596fc9fc415f59b93007e864bc I accidentally
flipped a condition that made the down key "recenter" on
the "normal" case, not in the "is-at-end" case.

This commit undoes that regression, which probably only
affected czo for a weekend, and makes the logic a bit
more clear.