pgroonga: Remove long-running update query in initial migration.
This query doesn't add any value, because it'll be overwritten in
migration 0002 anyway.  And because it isn't batched, it can take
several minutes to run on servers with hundreds of thousands to
millions of messages of history.  During that time, it's in a
transaction, and thus one can't send messages, so it forces downtime.